+++ public/js/controller/data/model/modelcontroller.js
@@ -18,10 +18,14 @@ class ModelController {
         if (apiModels) {
             var model;
             for (var data of apiModels) {
-                model = new XModel(data['definition']);
-                model.setId(data['id']);
-                model.init();
-                this._models.push(model);
+                try {
+                    model = new XModel(data['definition']);
+                    model.setId(data['id']);
+                    await model.initModel();
+                } catch (error) {
+                    app.controller.showError(error, "Initialising model '" + model.getName() + "' failed");
+                }
+                this._models.push(model); // still push to enable edit
             }
         }
         $(window).trigger('changed.model');
@@ -30,7 +34,7 @@ class ModelController {
 
     isModelDefined(name) {
         for (var model of this._models) {
-            if (model.getData().name === name)
+            if (model.getDefinition().name === name)
                 return true;
         }
         return false;
@@ -38,12 +42,12 @@ class ModelController {
 
     getModels(bIncludeSystem) {
         if (!bIncludeSystem)
-            return this._models.filter(function (x) { return x.getData()['name'].charAt(0) !== '_' });
+            return this._models.filter(function (x) { return x.getDefinition()['name'].charAt(0) !== '_' });
         else
             return this._models;
     }
 
     getModel(name) {
-        return this._models.filter(function (x) { return x.getData()['name'] === name })[0];
+        return this._models.filter(function (x) { return x.getDefinition()['name'] === name })[0];
     }
 }