+++ public/js/controller/data/route-controller.js
@@ -1,37 +1,44 @@
 class RouteController {
 
-    static CONFIG_ROUTES_IDENT = 'routes';
-
-    _url;
     _routes;
 
     constructor() {
     }
 
     async init() {
-        this._url = app.controller.getApiController().getApiOrigin() + "/api/_registry";
-        var entry = await WebClient.fetchJson(this._url + '?key=routes');
-        if (entry && entry.length == 1) {
-            var value = entry[0]['value'];
-            if (value)
-                this._routes = JSON.parse(value);
-        }
-        return Promise.resolve();
+        this._routes = [];
+        Promise.resolve();
     }
 
-    async setRoutes(routes) {
-        this._routes = routes;
-        return WebClient.request("PUT", this._url, { 'key': 'routes', 'value': JSON.stringify(this._routes) });
-    }
-
-    getRoutes() {
-        return this._routes;
-    }
-
-    getRoute(name) {
+    getMatchingRoute(path) {
         var res;
-        if (this._routes)
-            res = this._routes.filter(function (x) { return x.name === name })[0];
+        var match;
+        if (this._routes) {
+            for (var route of this._routes) {
+                if (route['regex'] && route['fn']) {
+                    match = new RegExp(route['regex'], 'ig').exec(path);
+                    if (match) {
+                        res = route;
+                        break;
+                    }
+                }
+            }
+        }
         return res;
+    }
+
+    addRoute(route) {
+        if (route['regex'] && route['fn']) {
+            this.deleteRoute(route);
+            this._routes.push(route);
+        }
+    }
+
+    deleteRoute(route) {
+        if (route['regex']) {
+            this._routes = this._routes.filter(function (x) {
+                x['regex'] !== route['regex'];
+            });
+        }
     }
 }